requirements.txt
click>=8.0
faiss-cpu>=1.7
gradio>=5.0
numpy>=1.24
openai>=1.0
sentence-transformers>=2.0
